+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes:
 Fixes:
 
   - Fix conversion from repeat pattern `a{n,m}` to ε-NFA correctly.
+  - Remove duplicated canonicalization of word escape class on `ignore-case`.
 
 # 0.2.0 (2020-11-01)